Deprecate old emmet
